Remove direct inspector wirings
The wirings have been moved to
segment-inspector/packages/webext/src/broker/index.ts
