+++ .changeset/popular-kangaroos-wave.md
@@ -1,0 +1,5 @@
+---
+'@segment/analytics-next': patch
+---
+
+Remove direct wirings for Segment Inspector
